Experiments using adapted RBF and standard RBF:  - real-world data/experiments: abalone, boston housing, ankara weather,  auto_mpg  - simulation experiments using sinc signal.  - updated train_elm to fit with experiment setup.  - Included OLS training in the training options.
NOTE: I also added ailerone data to the experiment, but not sure if it's
a great idea to use it, since its too large and crashes my computer.
Provisions have been included to the code to make sure i don't
accidentally run it.

NOTE: I looked into implementing RBF using back-prop. It's not as
popular as I assumed, so I decided to let it go.

NS